Moved the float module and added cached powers for atof algorithm.

+++ src/atof/algorithm/decimal.rs
@@ -1,10 +1,10 @@
-//! Bigint implementation for
+//! Decimal (big integer) implementation for exact string-to-float conversions.
 
 // TODO(ahuszagh) Implement...
 
 /// Large, stack-allocated integer for exact string-to-float conversions.
 #[allow(dead_code)]     // TODO(ahuszagh) Remove
-pub(crate) struct BigInt {
+pub(crate) struct Decimal {
     /// Internal storage for the numbers.
     k: [u32; 40],
 }
